Fix the streak from not printing correctly. Just catch all exceptions when accessing the url cache

+++ src/main/java/net/bosselaar/seprinter/core/printer/ImageCreator.java
@@ -42,7 +42,7 @@ public class ImageCreator {
                     text2 = String.format("a sub to %s!", e.data.displayName);
                 } else {
                     text1 = String.format("%s subscribed!", e.data.displayName);
-                    text2 = String.format("%d month streak.", e.data.amount);
+                    text2 = String.format("%d month streak.", e.data.amount.longValue());
                 }
                 break;
